+++ app/services/export/conversions/rpa_sug_and_fa_letters_csv_export_service.rb
@@ -14,6 +14,9 @@ class Export::Conversions::RpaSugAndFaLettersCsvExportService < Export::CsvExpor
     school_main_contact_name
     school_main_contact_role
     school_main_contact_email
+    headteacher_contact_name
+    headteacher_contact_role
+    headteacher_contact_email
     chair_of_governors_name
     chair_of_governors_email
     academy_urn